+++ Game/AI/Decks/SkyStrikerExecutor.cs
@@ -159,7 +159,7 @@ namespace WindBot.Game.AI.Decks
             return base.OnPreBattleBetween(attacker, defender);
         }
 
-        public override bool OnSelectYesNo(int desc)
+        public override bool OnSelectYesNo(long desc)
         {
             if (desc == Util.GetStringId(CardId.SummonSorceress, 2)) // summon to the field of opponent?
                 return false;
