+++ Game/AI/Decks/ThunderDragonExecutor.cs
@@ -368,7 +368,7 @@ namespace WindBot.Game.AI.Decks
                 selectFlag[i] = false;
             }
         }
-        public override int OnSelectPlace(int cardId, int player, CardLocation location, int available)
+        public override int OnSelectPlace(long cardId, int player, CardLocation location, int available)
         {
             if (player == 0 && location==CardLocation.MonsterZone)
             {
@@ -415,7 +415,7 @@ namespace WindBot.Game.AI.Decks
             }
             return base.OnSelectPlace(cardId, player, location, available);
         }
-        public override IList<ClientCard> OnSelectCard(IList<ClientCard> cards, int min, int max, int hint, bool cancelable)
+        public override IList<ClientCard> OnSelectCard(IList<ClientCard> cards, int min, int max, long hint, bool cancelable)
         {
             if (selectFlag.Count(flag => flag) > 1)
             {
